# example-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для складов по городам (ДелаемIT)",
        "description": "Документация API для работы с городами и складами",
        "contact": {
            "email": "magoed694@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/city/{id}": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Получить информацию о городе",
                "operationId": "ce6226ed690386a2ace79830706623c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Информация о городе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Город не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Cities"
                ],
                "summary": "Обновить информацию о городе",
                "operationId": "ae2f61db36269d7e38de14019bf1de06",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Minsk"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Информация о городе обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Город не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cities"
                ],
                "summary": "Удалить город",
                "operationId": "0f2b6475aa2b18ca2828a5c571113f98",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Город удален"
                    },
                    "404": {
                        "description": "Город не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city": {
            "post": {
                "tags": [
                    "Cities"
                ],
                "summary": "Создать новый город",
                "operationId": "eb28b6babe5f520710176ba7677953a9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Minsk"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Город создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверный запрос"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/city/{id}/stocks": {
            "get": {
                "tags": [
                    "Cities"
                ],
                "summary": "Получить список складов города",
                "operationId": "cfbbb7b24d296144773c2ab6050cab7b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список складов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Stock"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Город не найден"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "City": {
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Minsk"
                    }
                },
                "type": "object"
            },
            "Stock": {
                "required": [
                    "id",
                    "city_id",
                    "name",
                    "lat",
                    "lng"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "city_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Main Warehouse"
                    },
                    "lat": {
                        "type": "numeric",
                        "example": "55,123456"
                    },
                    "lng": {
                        "type": "numeric",
                        "example": "23,123456"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Use Bearer Token for authorization",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}